chore: Break down elevation implementation to platform-specific files
